+++ src/components/basic_table/basic_table.tsx
@@ -884,7 +884,10 @@ export class EuiBasicTable<T = any> extends Component<
     const cells = [];
 
     const { itemId: itemIdCallback } = this.props;
-    const itemId: ItemIdResolved = getItemId(item, itemIdCallback) || rowIndex;
+    const itemId: ItemIdResolved =
+      getItemId(item, itemIdCallback) != null
+        ? getItemId(item, itemIdCallback)
+        : rowIndex;
     const selected = !selection
       ? false
       : this.state.selection &&